pyproject.toml
[tool.poetry]
name = "storage"
version = "0.1.0"
description = ""
authors = ["nez0xx <chugainov256@gmail.com>"]
readme = "README.md"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode="auto"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.2"}
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
aiosqlite = "^0.20.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
openpyxl = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
